drivers: pinctrl: Kconfig.imx: enable PINCTRL_IMX by default
Currently, when using the `pinctrl_imx.c` driver one has
to manually enable `CONFIG_PINCTRL_IMX` even if the pinctrl
node using the driver is enabled. This is redundant and prone
to error as one would expect the driver to be enabled when
the node is also enabled.

This commit fixes the issue by enabling `CONFIG_PINCTRL_IMX`
by default and adding a new dependency on the state of the
pinctrl node (i.e: node needs to be enabled for the configuration
to also be set to `y`).

(cherry picked from commit d42dfc36d7a88ba2fdb596a1cf0d2d24aab4815d)

Original-Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>
GitOrigin-RevId: d42dfc36d7a88ba2fdb596a1cf0d2d24aab4815d
Change-Id: I86a752cd44b59729f0ce52532f65abb3bc115776
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5438833
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>