drivers: serial: rpi_pico: Enable uart FIFO
A few lines above this change explicitly enable the UART FIFO.  Then this
line, goes and disables it.  This seems to be from a workaround where the
parity and stop bits are not coming in correctly in the config.  Fixing
that will be another patch.  But, the FIFO should always be enabled.  This
is visible when trying to use bit rates that are reasonably fast.  Even in
IRQ mode, with a sufficient bit rate, bytes are readily dropped.

Tested on a Sparkfun Promicro rp2040 board.

(cherry picked from commit f7e64c87032493c7c75562ea627b70649c0d18df)

Original-Signed-off-by: David Brown <david.brown@linaro.org>
GitOrigin-RevId: f7e64c87032493c7c75562ea627b70649c0d18df
Change-Id: I790a49bb5b373e38094c0382991f1649d859ab21
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5423756
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>