Bluetooth: BAP: Only dump active receive states
The bt_debug_dump_recv_state did not care if the receive
state being dumped to log was active or not, and if it was
not actice, then it may just contain random data.

Added a check to see if it was active.

(cherry picked from commit 16189b007f2c4930b00f56696f4ddeed96416899)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 16189b007f2c4930b00f56696f4ddeed96416899
Change-Id: Ia4c689378fa6d4561d3c1722ed868490a608c553
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5439941
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>