Bluetooth: Audio: Remove !found dbg log in codec.c
It is a common occurance that a specific type is not
found in the codec configuration, and the main way of
knowing that is by calling the function.

The debug log does not help debug anything (as it does also
return a specific error code), but rather just clutters up
the log if enabled.

(cherry picked from commit 1d5acb754727926bbe9fe701924d99da3a6eb102)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 1d5acb754727926bbe9fe701924d99da3a6eb102
Change-Id: Icf032d1ef12d0f810d78512964c2f839298b5738
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5438770
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>