doc: bsim: Add three more rules
On the request of the maintainer.

Background:

We have recently had issues with tests timing out because of the
difference of speed between CI runners and a standard developer machine.

In that case, the logs are printed, and some tests (by yours truly) are
a bit on the chatty side. Github's UI chokes (for a good minute or more
lol) on more than a few hundred lines, so it's a bad experience for
contributors.

(cherry picked from commit d73b6e50b9e74f6044e92f1f9f45b69f879121d1)

Original-Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>
GitOrigin-RevId: d73b6e50b9e74f6044e92f1f9f45b69f879121d1
Change-Id: I9128b7442ada25f8544bb291b269fd7de0cf0cb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5402698
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>