drivers: serial: pl011: Use clock control to determine frequency
If a clock control device is set in the `clocks` property,
it is used to determine the frequency.
If `fixed-clock` is set for the clocks, we use that value as is.

(cherry picked from commit dd84e8db26dfd7c08cad381f12794679abad3d94)

Original-Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@gmail.com>
GitOrigin-RevId: dd84e8db26dfd7c08cad381f12794679abad3d94
Change-Id: I71b76f4cf9b2b99b3438cf6b4b7ff345108f2206
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5443074
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>