xtensa: make it work with TLB misses during interrupt handling
If there are any TLB misses during interrupt handling,
the user, kernel and double exception vector will be triggered
for the miss and the DEPC and EXCCAUSE overwritten as the TLB
missse are be handled in the assembly code and execution
returned to the original vector code. Because of this, both
DEPC and EXCCAUSE being read in the C handler are not the ones
that triggered the original exception (for example, level-1
interrupt). So stash both DEPC and EXCCAUSE such that
the original cause of exception is visible in the C handler.

(cherry picked from commit bc3e77b356e29d2f3bb41e6e09ee064016a76318)

Original-Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: bc3e77b356e29d2f3bb41e6e09ee064016a76318
Change-Id: I2bacfbb4cba9fa3642425113bb7d5616a1a01c34
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5635435
Tested-by: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Patryk Duda <patrykd@google.com>
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>