dma: Fix error_callback enable/disable confusion
Previously the logic was inverted for error_callback_en where 0 was
enablement and 1 was disable. This was likely done so that the default,
sensibly so, was to enable the error callback if possible. A variety of
in tree users had confused the enable/disable value.

Change the name of the flag to error_callback_dis where the default
remains 0 (do not disable the callback!) and correct in tree uses of the
flag where it seemed incorrect.

(cherry picked from commit 4180d70439f599a3597b2e31753773cfa222ace2)

Original-Signed-off-by: Tom Burdick <thomas.burdick@intel.com>
GitOrigin-RevId: 4180d70439f599a3597b2e31753773cfa222ace2
Change-Id: I4ce177d7ccae1e311e967a26fd0d134583d4b8a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5449718
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>