drivers: video: csi: Fix get_fmt
The CSI should always propagate the get_fmt request down to the sensor.
It should not return a format by it own unless it contains a test
pattern generator inside it.

The problem with the legacy code is that this code snippet is never
reached if the sensor could return a format. In case the sensor failed
to return a format, these codes are reached but the returned format
makes no sense and its value was not initialized.

(cherry picked from commit 46eb7e2b65f799f3b81e000657cf2eb08522b600)

Original-Signed-off-by: Phi Bang Nguyen <phibang.nguyen@nxp.com>
GitOrigin-RevId: 46eb7e2b65f799f3b81e000657cf2eb08522b600
Change-Id: I65de88c7b45935c933516e066f43a328aacaf0de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5533413
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>