Bluetooth: Assert alignof bt_addr types
I got some feedback about confusion as to why the bt_addr types are not
packed. This commit adds a note about it.

This commit also adds a check for the alignof of the structs to verify
the statement 'their members are bytes or byte arrays' to thoroughly
prove that the struct is effectively as-if packed.

(cherry picked from commit 2ec7a46d0fb2730682ba25d5ca9d494000399ef8)

Original-Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>
GitOrigin-RevId: 2ec7a46d0fb2730682ba25d5ca9d494000399ef8
Change-Id: Ic624967898b3bcd10fc690001000fbcfaae04d68
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5531748
Tested-by: Ting Shen <phoenixshen@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>