xtensa: mmu: MMU re-initialization API
With power managment is enabled, depending on the SoC power state
used when idle, the MMU may lose context and may need to be re-initialized.
When re-initializing the MMU, we must not re-create the page table
because it may overwrite changes done during the execution, but we still
need to set the asid and page table for the current context.

(cherry picked from commit f74a84b251cd21a71ce0c6f47019919c68730d27)

Original-Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>
GitOrigin-RevId: f74a84b251cd21a71ce0c6f47019919c68730d27
Change-Id: I705e4058c373143c7bea576f4e1a7ddc1e9c760b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5600972
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>