sysbuild: cmake: fix sysbuild_cache_set() function
The sysbuild_cache_set() supports setting variables in a dedicated
sysbuild image cache. When appending to a list, the list is checked for
existance of the variable to avoid appending it again on reruns.

This was done be checking with `string(FIND ...)`, but that will also
return an index on a partial match.

A much safer approach is to use the `if(<val> IN_LIST <list>)` approach
to check for existance.

(cherry picked from commit de53c0d2db19cc8d7ea55f27b6798f0c0c31574c)

Original-Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>
GitOrigin-RevId: de53c0d2db19cc8d7ea55f27b6798f0c0c31574c
Change-Id: I1a19fc62b9c5a70aee452436162216d7efd937af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5501559
Reviewed-by: Eric Yilun Lin <yllin@google.com>
Tested-by: Eric Yilun Lin <yllin@google.com>
Commit-Queue: Eric Yilun Lin <yllin@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>