Bluetooth: Remove rx < tx prio check
The comment states the reason for this requirement is to detect
violations by a remote peer. The ATT version of this was removed in
ea04fd95f939b8115358560df5887bb5f2baeab8. TLDR: It's neither possible
for, nor the job of the host to police the remote device.

We remove this requirement to be more flexible about the number of
priority levels in the system, and to avoid the temptation of using
priorities as a synchronization mechanism.

(cherry picked from commit 5dd1cdda3b69f0970466cc4ab232b2cd99dfa247)

Original-Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>
GitOrigin-RevId: 5dd1cdda3b69f0970466cc4ab232b2cd99dfa247
Change-Id: I754866d5e49c0c19a9099c489b46ed1a5c178d37
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5465674
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>