scripts: west: runner: openocd: Add gdb-client-port
The gdb-port defines the GDB port and the openocd runner passes
this value to openocd as well as gdb. However, the TI AM62x board
provides multiple ports for each subsystem. For example, systick
appears at 3333, A53 as 3334-3337, R5F as 3338 and M4F as 3339.

If we want to connect to the M4F, we need to add another port which
is different to the to the gdb-port value.

This patch adds an additional argument --gdb-client-port to define
the port which GDB should connect to. It defaults to 3333, identical
to gdb-port.

(cherry picked from commit b22c961e5e9caa7628630bfce1b81b2dfd15f6fd)

Original-Signed-off-by: Daniel Schultz <d.schultz@phytec.de>
GitOrigin-RevId: b22c961e5e9caa7628630bfce1b81b2dfd15f6fd
Change-Id: I9548bc539713c8f18f4bd6112189e8e5d3d44587
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5633770
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>