llext: set integration platform on relocatable test
This particular test was ignored in CI because, when using
platform_key, Twister picks the specific platform very early in
the process and it is possible that the chosen one is later
filtered out. To fix this, use the `integration_platforms` key
instead to specify the platforms where the test is expected to
complete successfully.

(cherry picked from commit 51195f48842d412967de602beccc530763c95808)

Original-Signed-off-by: Luca Burelli <l.burelli@arduino.cc>
GitOrigin-RevId: 51195f48842d412967de602beccc530763c95808
Change-Id: Iee005536a45d6fccd18bf13feeca0cc7d8f19525
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5468054
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>