Bluetooth: Tests: Remove usage of auto name in AD
The advertiser options to automatically add the name in the scan
response and the advertising data will be deprecated.

Update the tests that were relying on those options. A lot of test were
actually not using the name, for those the name have simply been
removed. For the others that were using the name, add it in the
advertising data or the scan response data.

(cherry picked from commit 9400823dcdd1ff98292eea05136f3afb363bb209)

Original-Signed-off-by: Théo Battrel <theo.battrel@nordicsemi.no>
GitOrigin-RevId: 9400823dcdd1ff98292eea05136f3afb363bb209
Change-Id: I7b2efad066de9015d43cf8a7e5fcca985ae92fa7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5480520
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>