net: if: Do not add IPv6 address lifetime timer twice to list
The IPv6 address lifetime timers are kept in a slist, but
the code that adds the entry to the list does not check
whether the item is already in the list. This will cause
problems when trying to remove the address from the list.

Normally this is not causing issues, but if the function
net_if_ipv6_addr_update_lifetime() is called multiple times
before the address expires, then the item was added to the
slist multiple times.

(cherry picked from commit b864cd9053a00032b993ac7e22fd34667850e419)

Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: b864cd9053a00032b993ac7e22fd34667850e419
Change-Id: I9f04262465b8cfbfc1423bfefa37b4065a51c9fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5469799
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>