release note/migration guide: Mention fuzzing changes
The fuzzing support only worked when targetting the
native_posix board, and in a way which was very coupled
with the fuzzing sample.
This has now been refactored, so those who may have
relied on hooking to the board in the same way as this
sample will need to adapt.

(cherry picked from commit 82fa83048e662602862292c741f15d049720fc96)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 82fa83048e662602862292c741f15d049720fc96
Change-Id: Iec30a053b6208af2cdd0f210b2359d098f966cfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5515346
Tested-by: Eric Yilun Lin <yllin@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Eric Yilun Lin <yllin@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>