doc: posix: aep: condense aep documentation
The POSIX Application Environment Profiles are additive and
offer incrementally more features over the System Interfaces.

For example, PSE51 includes all of the features of the Minimal
Realtime System Profile plus all of the System Interfaces.
PSE52 includes all of the features of the Realtime Controller
System Profile, plus the features of PSE51, plus the System
Interfaces, and so on.

It makes sense to take advantage of that and to condense the
AEP documenation.

(cherry picked from commit c0dc8d14269dcd193192017e2daea72205da9f46)

Original-Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>
GitOrigin-RevId: c0dc8d14269dcd193192017e2daea72205da9f46
Change-Id: I62af9102cc4200f035ca17bb3a93ef754c9acbfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5501555
Tested-by: Eric Yilun Lin <yllin@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Eric Yilun Lin <yllin@google.com>
Commit-Queue: Eric Yilun Lin <yllin@google.com>