kernel: add CONFIG_DEVICE_DT_METADATA and helpers
This option allows you to look up a struct device from any of the
node labels that were attached to the devicetree node used to create
the device, etc.

This is helpful because node labels are a much more human-friendly set
of unique identifiers than the node names we are currently relying on
for use with device_get_binding(). Adding this infrastructure in the
device core allows anyone to make use of it without having to
replicate node label storage and search functions in various places in
the tree. The main use case, however, is for looking up devices by
node label in the shell.

Since there is a footprint penalty associated with storing the node
label metadata, leave this option disabled by default.

(cherry picked from commit 15c9d3711ced5f21474a2c68470fd7689e17dc87)

Original-Signed-off-by: Martí Bolívar <mbolivar@amperecomputing.com>
GitOrigin-RevId: 15c9d3711ced5f21474a2c68470fd7689e17dc87
Change-Id: Ia641cefe412fc7772b3fdefe4ec2db8445ec1432
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5629313
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>