ARC: fault: Fix uninitialized memory access
Found via static analysis. In fault path when checking for stack
overflows, if CONFIG_MULTITHREADING is not set, `guard_end` is left
uninitialized and is subsequently used in a comparison.

The solution is to simply return `false` in this configuration as stack
guards are not configured in the first place.

(cherry picked from commit 3f78ca987333810f24dcc936beeb601f9dbd7a3f)

Original-Signed-off-by: François Baldassari <francois@memfault.com>
GitOrigin-RevId: 3f78ca987333810f24dcc936beeb601f9dbd7a3f
Change-Id: I73c1dfb9aee64703b00ae6e693b06d794e13ddef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5501573
Reviewed-by: Eric Yilun Lin <yllin@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Eric Yilun Lin <yllin@google.com>
Commit-Queue: Eric Yilun Lin <yllin@google.com>