zbus: vded: msg_sub: Improve channel ref copy
The VDED was adding the channel information per clone. But it could be
done in the original buffer. This commit fixes that by adding the
channel information to the original buffer and not for each clone. As a
result, we have a more straightforward VDED execution with fewer copies.

(cherry picked from commit 453ab8a9a356acf475a965a777a370795effa255)

Original-Signed-off-by: Rodrigo Peixoto <rodrigopex@gmail.com>
GitOrigin-RevId: 453ab8a9a356acf475a965a777a370795effa255
Change-Id: Iebc74c1029c7d446326d10d5b54a1f2ac4d8042e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5613850
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>