mbedtls: introduce PSA-specific Kconfig file
It is meant specifically for configuration of the PSA crypto library.

The underlying PSA configuration items are guarded by the condition
that a PSA crypto provider must be present, which is the case when
either TF-M is in use or MbedTLS's PSA core is built as part of
the application image.

(cherry picked from commit 2d61db9a606d155c2c2b6d04f8f5db436cc33483)

Original-Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>
GitOrigin-RevId: 2d61db9a606d155c2c2b6d04f8f5db436cc33483
Change-Id: I85e5e9a8644b76acadd899182a5b4bf79ae2230f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5533416
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>